Preserve the next hop on BGP advertisements
When #206 was merged to fix #199, this had the unintended consequence
of also causing nodes not acting as route reflectors to re-advertise
static routes learnt via node-to-node mesh. This meant all nodes
were receiving traffic for services with externalTrafficPolicy: Local
even if they were not running an endpoint for the service.

This change ensures that the correct next hop is advertised to any
bgp peers.